==> 03_quality_scorer.ipynb <==
# Argument Quality Scorer

## Import packages

import pandas as pd
import torch
from argsum.tools import train_quality_scorer, eval_quality_scorer

## Load datasets

# Load ArgKP-2021 
ArgKP21 = pd.read_csv('data/ArgKP-2021/dataset_splits_scores.csv')
ArgKP21_train_topics = ArgKP21[ArgKP21['set'] == 'train']['topic'].unique()
ArgKP21_dev_topics = ArgKP21[ArgKP21['set'] == 'dev']['topic'].unique()
ArgKP21_test_topics = ArgKP21[ArgKP21['set'] == 'test']['topic'].unique()

# Load ArgQ dataset
ArgQ = pd.read_csv('data/IBM-ArgQ-Rank-30kArgs/dataset_scores.csv')
ArgQ_train = ArgQ[ArgQ['topic'].isin(ArgKP21_train_topics)]
ArgQ_dev = ArgQ[ArgQ['topic'].isin(ArgKP21_dev_topics)]

test_set_mask = (ArgQ['set'] == 'test') & (~ArgQ['topic'].isin(ArgKP21_train_topics.tolist() + ArgKP21_dev_topics.tolist()))
ArgQ_test = ArgQ[test_set_mask]

## Train, development and test topics

print('Train Topics:\n')

for topic in ArgQ_train['topic'].unique():
    print(topic)

print('Train Topics:\n')

for topic in ArgQ_dev['topic'].unique():
    print(topic)

print('Train Topics:\n')

for topic in ArgQ_test['topic'].unique():
    print(topic)

set(ArgQ_train['topic'].unique()).intersection(set(ArgQ_dev['topic'].unique())).intersection(set(ArgQ_test['topic'].unique()))

## Train model for different settings

for p in [True, False]:
    for t in ['MACE-P', 'WA']:
        train_quality_scorer(ArgQ_train, ArgQ_dev, target_name = t, pooling = p)

## Compare model performances

# Set model directories 
model_dirs = ['models/quality_scorer/bert_ft_topic_np_mace-p/2024-May-29_15-58-03',
              'models/quality_scorer/bert_ft_topic_np_wa/2024-May-29_16-10-44',
              'models/quality_scorer/bert_ft_topic_p_mace-p/2024-May-29_15-30-54',
              'models/quality_scorer/bert_ft_topic_p_wa/2024-May-29_15-43-23']
# Evaluate models 
for dir in model_dirs:
    model = torch.load(dir + '/best_model.pt')
    eval_quality_scorer(model, ArgQ_test, output_dir = dir)

# Evaluate debater api
eval_quality_scorer('debater_api', ArgQ_test, output_dir = 'models/quality_scorer/debater_api')

model_dirs = ['models/quality_scorer/bert_ft_topic_np_mace-p/2024-May-29_15-58-03',
              'models/quality_scorer/bert_ft_topic_np_wa/2024-May-29_16-10-44',
              'models/quality_scorer/bert_ft_topic_p_mace-p/2024-May-29_15-30-54',
              'models/quality_scorer/bert_ft_topic_p_wa/2024-May-29_15-43-23']

# Load evaluation reports
np_macep = pd.read_csv(model_dirs[0] + '/eval_report.csv')
np_wa = pd.read_csv(model_dirs[1] + '/eval_report.csv')
p_macep = pd.read_csv(model_dirs[2] + '/eval_report.csv')
p_wa = pd.read_csv(model_dirs[3] + '/eval_report.csv')
debater_api = pd.read_csv('models/quality_scorer/debater_api/eval_report.csv')
# Set model names
model_names = ['np_mace-p', 'np_wa', 'p_mace-p', 'p_wa', 'debater_api']
# Print concatenated evaluation reports
results = pd.concat([np_macep, np_wa, p_macep, p_wa, debater_api])
results.index = model_names
results.round(3)
