==> 04_match_scorer.ipynb <==
# Argument Match Scorer

## Import packages

import pandas as pd
import numpy as np
import torch
from sentence_transformers import SentenceTransformer
from argsum import train_cross_match_scorer, train_bi_match_scorer_smatchtopr, train_bi_match_scorer_barh, eval_match_scorer
from argsum import load_test_df

## Load datasets

# Load ArgKP-2021 
ArgKP21 = pd.read_csv('data/ArgKP-2021/dataset_splits_scores.csv')
ArgKP21_train = ArgKP21[ArgKP21['set'] == 'train']
ArgKP21_dev = ArgKP21[ArgKP21['set'] == 'dev']
ArgKP21_test_processed = load_test_df('ArgKP21')

## Train cross-encoder match scorer

###  Without pooling layer

**RoBERTa**

model_path = 'roberta-large'
train_cross_match_scorer(model_path, ArgKP21_train, ArgKP21_dev, seed = 384, 
                         pooling = False, include_topic = False, augmented_data = False)

cross_roberta_np = torch.load('models/match_scorer/cross_encoder/roberta_np/2024-Feb-20_08-20-28/best_model.pt')
cross_roberta_np_results = eval_match_scorer(model = cross_roberta_np, df_test = ArgKP21_test_processed, include_topic = False, output_dir = None)
cross_roberta_np_results.round(3)

**DeBERTa**

[1] [Welcome to the Real World: Efficient, Incremental and Scalable Key Point Analysis](https://aclanthology.org/2023.emnlp-industry.46.pdf) (Eden et al., 2023)

- Lower learning rate than described in [1] (5e-6) because of overfitting

model_path = 'microsoft/deberta-v3-large'
train_cross_match_scorer(model_path, ArgKP21_train, ArgKP21_dev, num_train_epochs = 6, learning_rate = 2e-6 ,seed = 467, 
                         pooling = False, include_topic = False, augmented_data = False)

cross_deberta_np = torch.load('models/match_scorer/cross_encoder/deberta-v2_np/2024-Feb-20_13-41-02/best_model.pt')
cross_deberta_np_results = eval_match_scorer(model = cross_deberta_np, df_test = ArgKP21_test_processed, include_topic = False, output_dir = None)
cross_deberta_np_results.round(3)

model_path = 'deberta-v3-large'
train_cross_match_scorer(model_path, ArgKP21_train, ArgKP21_dev, num_train_epochs = 6, learning_rate = 5e-6 ,seed = 467, 
                         pooling = False, include_topic = False, augmented_data = False)

cross_deberta_np = torch.load('models/match_scorer/cross_encoder/deberta-v2_np/2024-Feb-20_13-41-02/best_model.pt')
cross_deberta_np_results = eval_match_scorer(model = cross_deberta_np, df_test = ArgKP21_test_processed, include_topic = False, output_dir = None)
cross_deberta_np_results.round(3)

### With pooling layer

**RoBERTa**

model_path = 'roberta-large'
train_cross_match_scorer(model_path, ArgKP21_train, ArgKP21_dev, seed = 384, 
                         pooling = True, include_topic = False, augmented_data = False)

cross_roberta_p = torch.load('models/match_scorer/cross_encoder/roberta_p/2024-Feb-20_10-20-20/best_model.pt')
cross_roberta_p_results = eval_match_scorer(model = cross_roberta_p, df_test = ArgKP21_test_processed, include_topic = False, output_dir = None)
cross_roberta_p_results.round(3)

**DeBERTa**

model_path = 'microsoft/deberta-v3-large'
train_cross_match_scorer(model_path, ArgKP21_train, ArgKP21_dev, num_train_epochs = 6, learning_rate = 2e-6 ,seed = 467, 
                         pooling = True, include_topic = False, augmented_data = False)

cross_deberta_p = torch.load('models/match_scorer/cross_encoder/deberta-v2_p/2024-Feb-20_15-30-58/best_model.pt')
cross_deberta_p_results = eval_match_scorer(model = cross_deberta_p, df_test = ArgKP21_test_processed, include_topic = False, output_dir = None)
cross_deberta_p_results.round(3)

## Train bi-encoder match scorer

### RoBERTa with topic information (SMatchToPr)

train_bi_match_scorer_smatchtopr(ArgKP21_train, include_topic = True)

bi_roberta = SentenceTransformer(f'models/match_scorer/bi_encoder/roberta_tp/2024-Feb-20_16-23-49')
bi_roberta_results = eval_match_scorer(model = bi_roberta, df_test = ArgKP21_test_processed, include_topic = True, output_dir = None)
bi_roberta_results.round(3)

bi_roberta = SentenceTransformer(f'models/match_scorer/bi_encoder/roberta_tp/2024-Feb-20_16-23-49')
bi_roberta_results = eval_match_scorer(model = bi_roberta, df_test = ArgKP21_test_processed, include_topic = False, output_dir = None)
bi_roberta_results.round(3)

### all-mpnet-base-v2 (own)

train_bi_match_scorer_barh(ArgKP21_train, loss = 'cos')

bi_roberta = SentenceTransformer(f'models/match_scorer/bi_encoder/all-mpnet-base-v2/2024-Jun-05_22-29-27')
bi_roberta_results = eval_match_scorer(model = bi_roberta, df_test = ArgKP21_test_processed, include_topic = False, output_dir = None)
bi_roberta_results.round(3)
