==> 05_flan_t5_sum.ipynb <==
# Flan T5 Summarization

https://github.com/HarrywillDr/KeyPoint-Analysis/blob/main/KPG/kpa_t5.py

## Import packages

import pandas as pd
from argsum import load_kp_test, preprocess_input_t5, train_t5, get_t5_cluster_sums

## Load data

# Load ArgKP21
df = pd.read_csv('data/ArgKP-2021/dataset_splits.csv')
# Load ArgKP21 train/dev splits
df_train = df[df['set'] == 'train']
df_dev = df[df['set'] == 'dev']
# Load KP_test
KP_test = load_kp_test()
# Preprocess ArgKP21 train/dev splits
df_train = preprocess_input_t5(df_train)
df_dev = preprocess_input_t5(df_dev)

train_t5(df_train, df_dev, learning_rate = 4e-4, load_best_model_at_end = False)

## Inspect summaries of fine-tuned T5

topic = KP_test['topic'].unique()[0]
stance = KP_test['stance'].unique()[1]
kps = KP_test[(KP_test['topic'] == topic) & (KP_test['stance'] == stance)]['key_point'].to_list()
kps_unique = set(list(kps))
args = KP_test[(KP_test['topic'] == topic) & (KP_test['stance'] == stance)]['argument'].to_list()
kps_unique_dict = dict(zip(kps_unique, [i for i in range(len(kps_unique))]))
gold_cluster_ids = [kps_unique_dict[kps[i]] for i in range(len(kps))]

kps_unique

sums = get_t5_cluster_sums(args, gold_cluster_ids, topic, stance, num_beams = 6, n = 5)
sums
